requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
plotly>=5.18.0
openpyxl>=3.1.0

# Google Sheets support (optional)
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.70.0
